Update button documentation to reference disabled state
Following a discussion arising from the use of disabled buttons in
WorldPay pages, we discussed the potential to reference them in the
design system - primarily for the benefit of third-parties.

WorldPay hosted payment pages[^1] link the disabled state of
several buttons to the state of other fields on the page (for example,
a "Make payment" button will be disabled until all required fields
have been populated with valid entries).

We are not keen on the use of dynamically updating the disabled state
and prefer server-side validation and a consistent user interface for
users as they interact with forms.

The content introduced borrows from the GOV.UK Design System [^2] but:

1. Clarifies that disabled buttons have been intentionally omitted
2. Omits the GDS reference to poor contrast (because that is not the
   case for disabled buttons using our Design System styles)

[^1]: https://developer.worldpay.com/products/access/hosted-payment-pages
[^2]: https://design-system.service.gov.uk/components/button/#disabled-buttons